feat: id validator

+++ build/infrastructures/repository/index.js
@@ -8,6 +8,10 @@ var _regenerator = _interopRequireDefault(require("@babel/runtime/regenerator"))
 
 var _asyncToGenerator2 = _interopRequireDefault(require("@babel/runtime/helpers/asyncToGenerator"));
 
+var _error = _interopRequireDefault(require("../../module/error"));
+
+var _isValidId = _interopRequireDefault(require("../../module/isValidId"));
+
 function ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }
 
 function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(Object(source), true).forEach(function (key) { (0, _defineProperty2["default"])(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }
@@ -41,11 +45,29 @@ module.exports = function (Model) {
   }();
 
   var update = function update(data) {
-    return Model.findByIdAndUpdate(data.id, _objectSpread({}, data)).exec();
+    try {
+      var id = data.id;
+
+      if (!(0, _isValidId["default"])(id)) {
+        throw _error["default"].ValidationError("Invalid Id");
+      }
+
+      return Model.findByIdAndUpdate(id, _objectSpread({}, data)).exec();
+    } catch (error) {
+      throw error;
+    }
   };
 
   var remove = function remove(id) {
-    return Model.findByIdAndRemove(id).exec();
+    try {
+      if (!(0, _isValidId["default"])(id)) {
+        throw _error["default"].ValidationError("Invalid Id");
+      }
+
+      return Model.findByIdAndRemove(id).exec();
+    } catch (error) {
+      throw error;
+    }
   };
 
   var getProducts = function getProducts(_ref2) {
@@ -55,7 +77,15 @@ module.exports = function (Model) {
   };
 
   var findById = function findById(id) {
-    return Model.findById(id).exec();
+    try {
+      if (!(0, _isValidId["default"])(id)) {
+        throw _error["default"].ValidationError("Invalid Id");
+      }
+
+      return Model.findById(id).exec();
+    } catch (error) {
+      throw error;
+    }
   };
 
   var findBy = function findBy(query) {
@@ -97,7 +127,7 @@ module.exports = function (Model) {
                 break;
               }
 
-              throw Error("Category ".concat(categories[i], " doesn't exist"));
+              throw _error["default"].NotFoundError("Category ".concat(categories[i], " doesn't exist"));
 
             case 8:
               i++;
